+++ packages/o-is-react/src/condition/empty.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import FieldPicker from './field-picker'
+import FieldPicker from '../field-picker'
 
 class EmptyCondition extends React.Component {
